Add AstroCateogry meta with autocomplete feature
Added the use of comboboxcontrol with dynamic fetching of the categiry using the input typed by the editor. This can replace the normal dropdown feature if we have large amount of the data which look odd while using dropdown. It is clean and have a good UI for editor with great experience.